fix: 테스트케이스 선택 로직 이동
