+++ app/src/pages/spaces/projects/testruns/TestcaseSelectPopup/TestcaseSelectPopup.jsx
@@ -56,138 +56,6 @@ function TestcaseSelectPopup({ testcaseGroups, selectedTestcaseGroups, setOpened
     setCurrentSelectedTestcaseGroups(cloneDeep(selectedTestcaseGroups));
   }, [selectedTestcaseGroups]);
 
-  const removeParentId = (list, parentId) => {
-    const targetGroupIds = testcaseGroups.filter(d => d.parentId === parentId).map(d => d.id);
-
-    for (let i = list.length - 1; i >= 0; i -= 1) {
-      if (targetGroupIds.includes(list[i].testcaseGroupId)) {
-        list.splice(i, 1);
-      }
-    }
-
-    for (let i = 0; i < targetGroupIds.length; i += 1) {
-      removeParentId(list, targetGroupIds[i]);
-    }
-  };
-
-  const addChildGroup = (list, parentId) => {
-    const targetGroupIds = testcaseGroups.filter(d => d.parentId === parentId).map(d => d.id);
-    for (let i = 0; i < targetGroupIds.length; i += 1) {
-      if (!list.find(d => d.testcaseGroupId === targetGroupIds[i])) {
-        const testcaseGroupInfo = testcaseGroups.find(d => d.id === targetGroupIds[i]);
-        list.push({
-          testcaseGroupId: targetGroupIds[i],
-          testcases:
-            testcaseGroupInfo.testcases?.map(d => {
-              return {
-                testcaseId: d.id,
-              };
-            }) || [],
-        });
-      }
-    }
-
-    for (let i = 0; i < targetGroupIds.length; i += 1) {
-      addChildGroup(list, targetGroupIds[i]);
-    }
-  };
-
-  const addParentGroup = (list, groupId) => {
-    const groupInfo = testcaseGroups.find(d => d.id === groupId);
-    if (groupInfo.parentId) {
-      const parentGroupId = testcaseGroups.find(d => d.id === groupInfo.parentId).id;
-
-      if (!list.find(d => d.testcaseGroupId === parentGroupId)) {
-        list.push({
-          testcaseGroupId: parentGroupId,
-          testcases: [],
-        });
-      }
-
-      addParentGroup(list, parentGroupId);
-    }
-  };
-
-  const onClick = (testcaseGroupId, testcaseId) => {
-    const isGroup = !testcaseId;
-
-    const nextCurrentSelectedTestcaseGroups = (currentSelectedTestcaseGroups || []).slice(0);
-
-    const selectedGroupIndex = nextCurrentSelectedTestcaseGroups.findIndex(d => d.testcaseGroupId === testcaseGroupId);
-
-    if (isGroup) {
-      if (selectedGroupIndex > -1) {
-        const testcaseGroupInfo = testcaseGroups.find(d => d.id === testcaseGroupId);
-        nextCurrentSelectedTestcaseGroups.splice(selectedGroupIndex, 1);
-        removeParentId(nextCurrentSelectedTestcaseGroups, testcaseGroupInfo.id);
-      } else {
-        const testcaseGroupInfo = testcaseGroups.find(d => d.id === testcaseGroupId);
-
-        let parentGroupId = testcaseGroupInfo.parentId;
-        while (parentGroupId) {
-          if (!nextCurrentSelectedTestcaseGroups.find(d => d.testcaseGroupId === parentGroupId)) {
-            nextCurrentSelectedTestcaseGroups.push({
-              testcaseGroupId: parentGroupId,
-              testcases: [],
-            });
-          }
-
-          const parentGroup = testcaseGroups.find(d => d.id === parentGroupId);
-          parentGroupId = parentGroup?.parentId;
-        }
-
-        nextCurrentSelectedTestcaseGroups.push({
-          testcaseGroupId,
-          testcases:
-            testcaseGroupInfo.testcases?.map(d => {
-              return {
-                testcaseId: d.id,
-              };
-            }) || [],
-        });
-
-        addChildGroup(nextCurrentSelectedTestcaseGroups, testcaseGroupId);
-      }
-    } else if (selectedGroupIndex > -1) {
-      const currentSelectedInfo = nextCurrentSelectedTestcaseGroups[selectedGroupIndex];
-      const selectedTestcaseIndex = currentSelectedInfo.testcases?.findIndex(d => d.testcaseId === testcaseId);
-      if (selectedTestcaseIndex > -1) {
-        currentSelectedInfo.testcases.splice(selectedTestcaseIndex, 1);
-      } else {
-        currentSelectedInfo.testcases.push({
-          testcaseId,
-        });
-      }
-    } else {
-      nextCurrentSelectedTestcaseGroups.push({
-        testcaseGroupId,
-        testcases: [
-          {
-            testcaseId,
-          },
-        ],
-      });
-
-      addParentGroup(nextCurrentSelectedTestcaseGroups, testcaseGroupId);
-    }
-
-    setCurrentSelectedTestcaseGroups(nextCurrentSelectedTestcaseGroups);
-  };
-
-  const filteredProjectTestcaseGroupTree = useMemo(() => {
-    return projectTestcaseGroupTree
-      ?.filter(testcaseGroup => testcaseUtil.isGroupFilteredByName(testcaseGroup, filterCondition.name))
-      ?.filter(testcaseGroup => testcaseUtil.isGroupFilteredByRange(testcaseGroup, filterCondition.minDate, filterCondition.maxDate))
-      .map(testcaseGroup => {
-        return {
-          ...testcaseGroup,
-          testcases: testcaseGroup.testcases
-            .filter(testcase => testcaseUtil.isFilteredTestcaseByName(testcase, filterCondition.name))
-            .filter(testcase => testcaseUtil.isFilteredTestcaseByRange(testcase, filterCondition.minDate, filterCondition.maxDate)),
-        };
-      });
-  }, [projectTestcaseGroupTree, filterCondition]);
-
   const allCheck = useCallback(() => {
     if (currentSelectedTestcaseGroups.length > 0) {
       setCurrentSelectedTestcaseGroups([]);
@@ -311,10 +179,10 @@ function TestcaseSelectPopup({ testcaseGroups, selectedTestcaseGroups, setOpened
               />
             </div>
             <TestcaseSelector
-              filteredProjectTestcaseGroupTree={filteredProjectTestcaseGroupTree}
+              testcaseGroups={testcaseGroups}
               currentSelectedTestcaseGroups={currentSelectedTestcaseGroups}
               filterCondition={filterCondition}
-              onClick={onClick}
+              onChange={setCurrentSelectedTestcaseGroups}
             />
           </div>
         )}
